+++ tests/test_utils.py
@@ -1,9 +1,38 @@
 
 import time
+import random
+
 from mock import MagicMock
 
 from delivery.models.project import Project
 from delivery.models.runfolder import Runfolder
+from delivery.services.external_program_service import ExecutionResult, Execution
+
+class MockIOLoop():
+    def __init__(self):
+        pass
+
+    def spawn_callback(self, f, **args):
+        f(**args)
+
+class MockExternalRunnerService():
+
+    def __init__(self, return_status=0, throw=False):
+        self.return_status = return_status
+        self.throw = throw
+
+    def run(self, cmd):
+        if self.throw:
+            raise Exception("Test the exception handling...")
+        mock_process = MagicMock()
+        execution = Execution(pid=random.randint(1, 1000), process_obj=mock_process)
+        return execution
+
+    def wait_for_execution(self, execution):
+        time.sleep(0.1)
+        return ExecutionResult(status_code=self.return_status,
+                               stderr="stderr",
+                               stdout="stdout")
 
 
 class TestUtils:
